Add a few plot styling functions
This patch introduces the ability to select a font for the plot (font
must be already installed on the system), as well as a function that
prettifies the plot.
The prettify() function bascially sets a default value for some of
the most common settings people change about their plots. Obviously,
it allows the user to change some of the settings it sets as well.