Add ability to change title/x-y labels
In this patch, we added the ability to set the title and labels of the
plots using the modify_plot() function. Normally, we'd have to call
3 different functions on out plot in order to be able to customize its
title and labels, but the modify_plot() function will massively simplify
this process.

In addition, the src/ folder is now named plotfa/. This is purely for
beauty purposes.