Change farsi convertor name & signature
In this patch, I changed the farsi convertor's name from farsi() to fa()
because this makes more sense and leads to more elegant code. On top of
this, I improved how the fa() functions handles its input; now if we
want to convert a list of persian text, we have to supply a kwarg to
the function. This is far better than the previous version which had a
ton of dirty hacks in it.